Added documentation and tests for the apply-diff function which will be used in the context of change tracking.
Instead of making a system that looks for what has changed before and after a :db effect, we will have event handlers providing a diff description, and an effect handler which is going to apply it to the value of the db. This diff will then be used to find things which depend on the data being changed and update them.